+++ packages/core/src/components/ozon-content/nodes/int-ref.node.tsx
@@ -1,18 +1,31 @@
 import { h } from "@stencil/core";
 
+import { inputToXmlDocument } from "../functions/input-to-xml-document.function";
 import { OzonContentNodeContext } from "../ozon-content-node-context.interface";
 import { OzonContentNode } from "../ozon-content-node.interface";
 
 export class OzonContentIntRefNode implements OzonContentNode {
   name = "IntRef";
 
-  render(node: Element, { mapNodeToJsx, emitAnchorClick }: OzonContentNodeContext) {
-    const ref = node.getAttribute("ref");
-    if (!ref) {
-      return mapNodeToJsx(node.childNodes);
+  render(node: Element, { mapNodeToJsx, emitClick, begripResolver }: OzonContentNodeContext) {
+    const name = "IntRef" as const;
+    const scope = node.getAttribute("scope");
+    const value = node.getAttribute("ref");
+
+    if (scope === "Begrip") {
+      const begrip = begripResolver ? begripResolver(name, "ref", value, node) : value;
+
+      return (
+        begrip && (
+          <dso-ref-toggletip icon="info">
+            <span slot="label">{mapNodeToJsx(node.childNodes)}</span>
+            {mapNodeToJsx(inputToXmlDocument(begrip, new DOMParser()))}
+          </dso-ref-toggletip>
+        )
+      );
     }
 
-    const intRefOnClick = (event: MouseEvent) => {
+    const handleIntRefClick = (event: MouseEvent) => {
       event.preventDefault();
 
       const target = event.currentTarget;
@@ -20,18 +33,15 @@ export class OzonContentIntRefNode implements OzonContentNode {
         return;
       }
 
-      const { href } = target;
-
-      emitAnchorClick({
-        node: this.name,
-        href,
-        documentComponent: ref,
+      emitClick({
+        type: "IntRef",
+        node,
         originalEvent: event,
       });
     };
 
     return (
-      <a href={`#${ref}`} onClick={intRefOnClick}>
+      <a href={`#${value}`} onClick={handleIntRefClick}>
         {mapNodeToJsx(node.childNodes)}
       </a>
     );