+++ packages/dso-toolkit/src/components/document-component/document-component.args.ts
@@ -3,7 +3,7 @@ import { HandlerFunction } from "storybook/actions";
 import { ArgTypes } from "storybook/internal/types";
 
 import { argTypeAction, isOdd } from "../../storybook";
-import { OzonContentUrlResolver } from "../ozon-content/ozon-content.models.js";
+import { OzonContentBegripResolver, OzonContentUrlResolver } from "../ozon-content/ozon-content.models.js";
 
 import {
   DocumentComponent,
@@ -38,13 +38,15 @@ export interface DocumentComponentArgs {
   mark?: string;
   enableRecursiveToggle?: boolean;
   mode: DocumentComponentMode;
+  dsoOzonContentClick: HandlerFunction;
   dsoTableOfContentsClick: HandlerFunction;
   ozonContentUrlResolver?: OzonContentUrlResolver;
+  ozonContentBegripResolver?: OzonContentBegripResolver;
 }
 
 export const documentComponentArgs: Omit<
   DocumentComponentArgs,
-  "dsoAnnotationToggle" | "dsoToggle" | "dsoMarkItemHighlight" | "dsoTableOfContentsClick"
+  "dsoAnnotationToggle" | "dsoToggle" | "dsoMarkItemHighlight" | "dsoTableOfContentsClick" | "dsoOzonContentClick"
 > = {
   annotated: true,
   bevatOntwerpInformatie: true,
@@ -85,6 +87,7 @@ export const documentComponentArgTypes: ArgTypes<DocumentComponentArgs> = {
   dsoAnnotationToggle: argTypeAction(),
   dsoToggle: argTypeAction(),
   dsoMarkItemHighlight: argTypeAction(),
+  dsoOzonContentClick: argTypeAction(),
   dsoTableOfContentsClick: argTypeAction(),
   filtered: {
     control: {
@@ -207,6 +210,13 @@ export function documentComponentMapper<TemplateFnReturnType>(
       }
 
       a.dsoTableOfContentsClick(e.detail);
+    },
+    dsoOzonContentClick: (e) => {
+      if (!e.detail.isModifiedEvent) {
+        e.detail.originalEvent.preventDefault();
+      }
+
+      a.dsoOzonContentClick(e.detail);
     },
     dsoToggle: (e) => a.dsoToggle(e.detail),
     children: a.open || a.openAnnotation ? children : undefined,