+++ packages/dso-toolkit/src/components/document-component/document-component.models.ts
@@ -1,4 +1,8 @@
-import { OzonContentText, OzonContentUrlResolver } from "../ozon-content/ozon-content.models.js";
+import {
+  OzonContentBegripResolver,
+  OzonContentText,
+  OzonContentUrlResolver,
+} from "../ozon-content/ozon-content.models.js";
 
 export interface DocumentComponent<TemplateFnReturnType> {
   alternativeTitle?: string;
@@ -26,7 +30,9 @@ export interface DocumentComponent<TemplateFnReturnType> {
   href?: string;
   dsoMarkItemHighlight?: (e: CustomEvent<DocumentComponentMarkItemHighlightEvent>) => void;
   dsoTableOfContentsClick?: (e: CustomEvent<DocumentComponentTableOfContentsClickEvent>) => void;
+  dsoOzonContentClick?: (e: CustomEvent<DocumentComponentOzonContentClickEvent>) => void;
   ozonContentUrlResolver?: OzonContentUrlResolver;
+  ozonContentBegripResolver?: OzonContentBegripResolver;
 }
 
 export type DocumentComponentHeading = "h2" | "h3" | "h4" | "h5" | "h6";
@@ -63,6 +69,30 @@ export interface DocumentComponentMarkItemHighlightEvent {
   source: DocumentComponentSource;
   text: string;
   elementRef: HTMLElement;
+}
+
+interface OzonContentClickBaseEvent<T extends string> {
+  type: T;
+  node: Element;
+  originalEvent: MouseEvent;
+}
+
+export type OzonContentClickKopEvent = OzonContentClickBaseEvent<"Kop">;
+export type OzonContentClickExtRefEvent = OzonContentClickBaseEvent<"ExtRef">;
+export type OzonContentClickIntIoRefToggleAnnotationEvent = OzonContentClickBaseEvent<"IntIoRefToggleAnnotation">;
+export type OzonContentClickIntRefEvent = OzonContentClickBaseEvent<"IntRef">;
+
+export type OzonContentClickEvent =
+  | OzonContentClickKopEvent
+  | OzonContentClickIntRefEvent
+  | OzonContentClickIntIoRefToggleAnnotationEvent
+  | OzonContentClickExtRefEvent;
+
+export interface DocumentComponentOzonContentClickEvent {
+  originalEvent: MouseEvent;
+  ozonContentClick: OzonContentClickEvent;
+  /** True when user clicked the card while holding Ctrl, Alt or other modifiers, or when the card is right-clicked. Can be used to determine navigation. */
+  isModifiedEvent: boolean;
 }
 
 export interface DocumentComponentTableOfContentsClickEvent {