+++ storybook/cypress/e2e/ozon-content.cy.ts
@@ -1,19 +1,39 @@
 import { isOdd } from "../support/is-odd";
 
 describe("Ozon Content", () => {
-  it("should emit anchorClick on IntRef anchor click", () => {
+  it("emits click with node='IntRef' on click of IntRef Link", () => {
     cy.visit("http://localhost:45000/iframe.html?id=core-ozon-content--int-ref");
 
     cy.injectAxe();
     cy.dsoCheckA11y("dso-ozon-content.hydrated");
 
     cy.get("dso-ozon-content.hydrated").then((c) => {
-      c.get(0).addEventListener("dsoAnchorClick", cy.stub().as("anchorClick"));
+      c.get(0).addEventListener("dsoClick", cy.stub().as("click"));
     });
 
-    cy.get("dso-ozon-content.hydrated").shadow().find("a[href = '#longTitle_inst2']").click();
+    cy.get("dso-ozon-content.hydrated").shadow().find("a[href = '#longTitle_inst2']").realClick();
 
-    cy.get("@anchorClick").should("have.been.calledOnce");
+    cy.get("@click")
+      .should("have.been.calledOnce")
+      .invoke("getCall", 0)
+      .its("args.0.detail.node")
+      .should("equal", "IntRef");
+
+    cy.get("dso-ozon-content.hydrated").matchImageSnapshot();
+  });
+
+  it('shows a toggletip on IntRef[@scope="begrip"]', () => {
+    cy.viewport(650, 650).visit("http://localhost:45000/iframe.html?id=core-ozon-content--int-ref-begrip");
+
+    cy.injectAxe();
+    cy.dsoCheckA11y("dso-ozon-content.hydrated");
+
+    cy.get("dso-ozon-content.hydrated")
+      .shadow()
+      .find("dso-ref-toggletip")
+      .shadow()
+      .find(".toggletip-button")
+      .realClick();
 
     cy.get("dso-ozon-content.hydrated").matchImageSnapshot();
   });
@@ -176,39 +196,64 @@ describe("Ozon Content", () => {
     cy.get("dso-ozon-content.hydrated").matchImageSnapshot();
   });
 
-  it("should render IntIoRef element", () => {
-    cy.visit("http://localhost:45000/iframe.html?id=core-ozon-content--int-io-ref");
+  describe("IntIoRef", () => {
+    beforeEach(() => {
+      cy.visit("http://localhost:45000/iframe.html?id=core-ozon-content--int-io-ref");
 
-    cy.injectAxe();
-    cy.dsoCheckA11y("dso-ozon-content.hydrated");
-
-    cy.get("dso-ozon-content.hydrated")
-      .shadow()
-      .find('a[href = "#gm0037_1__cmp_I__content_1__list_o_1__item_o_1__ref_o_1"]')
-      .should("exist");
-
-    cy.get("dso-ozon-content.hydrated").matchImageSnapshot();
-  });
-
-  it("should emit anchorClick on IntIoRef anchor click", () => {
-    cy.visit("http://localhost:45000/iframe.html?id=core-ozon-content--int-io-ref");
-
-    cy.get("dso-ozon-content.hydrated").then((c) => {
-      c.get(0).addEventListener("dsoAnchorClick", cy.stub().as("anchorClick"));
+      cy.get("dso-ozon-content.hydrated").then((c) => {
+        c.get(0).addEventListener("dsoClick", cy.stub().as("click"));
+      });
     });
 
-    cy.get("dso-ozon-content.hydrated")
-      .shadow()
-      .find('a[href = "#gm0037_1__cmp_I__content_1__list_o_1__item_o_1__ref_o_1"]')
-      .click();
+    it("shows a toggletip", () => {
+      cy.injectAxe();
+      cy.dsoCheckA11y("dso-ozon-content.hydrated");
 
-    cy.get("@anchorClick")
-      .should("have.been.calledOnce")
-      .invoke("getCall", 0)
-      .its("args.0.detail.node")
-      .should("equal", "IntIoRef");
+      cy.get("dso-ozon-content.hydrated")
+        .shadow()
+        .find("dso-ref-toggletip")
+        .shadow()
+        .find(".toggletip-button")
+        .realClick();
 
-    cy.get("dso-ozon-content.hydrated").matchImageSnapshot();
+      cy.get("dso-ozon-content.hydrated").matchImageSnapshot();
+    });
+
+    it("renders the 'Officiële publicaties' Link as ExtIoRef", () => {
+      cy.get("dso-ozon-content.hydrated")
+        .shadow()
+        .find("dso-ref-toggletip")
+        .shadow()
+        .find(".toggletip-button")
+        .realClick();
+
+      cy.get("dso-ozon-content.hydrated")
+        .shadow()
+        .find("dso-ref-toggletip")
+        .find("a")
+        .should("have.attr", "target", "_blank")
+        .and(
+          "have.attr",
+          "href",
+          "https://identifier-eto.overheid.nl//join/id/regdata/gm1979/2021/Delfzijlkamerverhuur/nld@2021-08-02;1",
+        )
+        .and("have.text", "Bedrijf categorie 2");
+    });
+
+    it("emits click with node='IntIoRefToggleAnnotation' on clicking the 'Gebieden op de kaart tonen' Button ", () => {
+      cy.get("dso-ozon-content.hydrated")
+        .shadow()
+        .find("dso-ref-toggletip")
+        .shadow()
+        .find(".toggletip-button")
+        .realClick();
+
+      cy.get("dso-ozon-content.hydrated").shadow().find("dso-ref-toggletip").find("button").click();
+
+      cy.get("@click").should("have.been.calledOnce");
+
+      cy.get("@click").invoke("getCall", 0).its("args.0.detail.node").should("equal", "IntIoRefToggleAnnotation");
+    });
   });
 
   it("should render ExtRef element", () => {